adapter/incus: split client and tests by domain
