+++ internal/adapter/incus/cluster.go
@@ -1,0 +1,154 @@
+package incus
+
+import (
+	"context"
+	"fmt"
+	"maps"
+	"net/url"
+
+	incusapi "github.com/lxc/incus/v7/shared/api"
+
+	"github.com/FuturFusion/operations-center/internal/provisioning"
+	"github.com/FuturFusion/operations-center/shared/api"
+)
+
+func (c Client) SetServerConfig(ctx context.Context, endpoint provisioning.Endpoint, config map[string]string) error {
+	client, err := c.getClient(ctx, endpoint)
+	if err != nil {
+		return err
+	}
+
+	svr, etag, err := client.GetServer()
+	if err != nil {
+		return fmt.Errorf("Failed to get current config from %q (%s): %w", endpoint.GetName(), endpoint.GetConnectionURL(), err)
+	}
+
+	if svr.Config == nil {
+		svr.Config = map[string]string{}
+	}
+
+	maps.Copy(svr.Config, config)
+
+	err = client.UpdateServer(svr.Writable(), etag)
+	if err != nil {
+		return fmt.Errorf("Failed to set config on %q (%s): %w", endpoint.GetName(), endpoint.GetConnectionURL(), err)
+	}
+
+	return nil
+}
+
+func (c Client) EnableCluster(ctx context.Context, server provisioning.Server) (clusterCertificate string, _ error) {
+	client, err := c.getClient(ctx, server)
+	if err != nil {
+		return "", err
+	}
+
+	req := incusapi.ClusterPut{
+		Cluster: incusapi.Cluster{
+			ServerName: server.Name,
+			Enabled:    true,
+		},
+	}
+
+	op, err := client.UpdateCluster(req, "")
+	if err != nil {
+		return "", fmt.Errorf("Failed to update cluster on %q (%s): %w", server.Name, server.GetConnectionURL(), err)
+	}
+
+	err = op.WaitContext(ctx)
+	if err != nil {
+		return "", fmt.Errorf("Failed to update cluster on %q (%s): %w", server.Name, server.GetConnectionURL(), err)
+	}
+
+	anyClusterCertificate, ok := op.Get().Metadata["certificate"]
+	if !ok {
+		return "", nil
+	}
+
+	clusterCertificate, ok = anyClusterCertificate.(string)
+	if !ok {
+		return "", nil
+	}
+
+	return clusterCertificate, nil
+}
+
+func (c Client) GetClusterNodeNames(ctx context.Context, endpoint provisioning.Endpoint) ([]string, error) {
+	client, err := c.getClient(ctx, endpoint)
+	if err != nil {
+		return nil, err
+	}
+
+	nodeNames, err := client.GetClusterMemberNames()
+	if err != nil {
+		return nil, fmt.Errorf("Failed to get cluster node names on %q (%s): %w", endpoint.GetName(), endpoint.GetConnectionURL(), err)
+	}
+
+	return nodeNames, nil
+}
+
+func (c Client) GetClusterJoinToken(ctx context.Context, endpoint provisioning.Endpoint, memberName string) (joinToken string, _ error) {
+	client, err := c.getClient(ctx, endpoint)
+	if err != nil {
+		return "", err
+	}
+
+	op, err := client.CreateClusterMember(incusapi.ClusterMembersPost{
+		ServerName: memberName,
+	})
+	if err != nil {
+		return "", fmt.Errorf("Failed to get cluster join token on %q (%s): %w", endpoint.GetName(), endpoint.GetConnectionURL(), err)
+	}
+
+	opAPI := op.Get()
+	token, err := opAPI.ToClusterJoinToken()
+	if err != nil {
+		return "", fmt.Errorf("Failed converting token operation to join token: %w", err)
+	}
+
+	return token.String(), nil
+}
+
+func (c Client) JoinCluster(ctx context.Context, server provisioning.Server, joinToken string, serverAddressOfClusterRole string, endpoint provisioning.Endpoint, config []api.ClusterMemberConfigKey) error {
+	client, err := c.getClient(ctx, server)
+	if err != nil {
+		return err
+	}
+
+	// Ignore error, connection URL has been parsed by incus client already.
+	clusterAddressURL, _ := url.Parse(endpoint.GetConnectionURL())
+
+	op, err := client.UpdateCluster(incusapi.ClusterPut{
+		Cluster: incusapi.Cluster{
+			ServerName:   server.Name,
+			Enabled:      true,
+			MemberConfig: config,
+		},
+		ClusterCertificate: endpoint.GetCertificate(),
+		ServerAddress:      serverAddressOfClusterRole,
+		ClusterToken:       joinToken,
+		ClusterAddress:     clusterAddressURL.Host,
+	}, "")
+	if err != nil {
+		return fmt.Errorf("Failed to update cluster during cluster join on %q (%s): %w", endpoint.GetName(), server.GetConnectionURL(), err)
+	}
+
+	err = op.WaitContext(ctx)
+	if err != nil {
+		return fmt.Errorf("Failed to wait for update operation during cluster join on %q (%s): %w", endpoint.GetName(), server.GetConnectionURL(), err)
+	}
+
+	return nil
+}
+
+func (c Client) UpdateClusterCertificate(ctx context.Context, endpoint provisioning.Endpoint, certificatePEM string, keyPEM string) error {
+	client, err := c.getClient(ctx, endpoint)
+	if err != nil {
+		return err
+	}
+
+	return client.UpdateClusterCertificate(incusapi.ClusterCertificatePut{
+		ClusterCertificate:    certificatePEM,
+		ClusterCertificateKey: keyPEM,
+	}, "")
+}
